+++ src/server/types.ts
@@ -226,6 +226,7 @@ export type LanguageServerOptions = {
 	configFile?: string;
 	customSyntax?: string;
 	ignoreDisables?: boolean;
+	ignorePath?: string;
 	packageManager: PackageManager;
 	reportInvalidScopeDisables?: boolean;
 	reportNeedlessDisables?: boolean;