+++ test/e2e/workspace/ignore-path/stylelint.config.js
@@ -1,0 +1,10 @@
+'use strict';
+
+/** @type {import('stylelint').Config} */
+const config = {
+	rules: {
+		indentation: [4],
+	},
+};
+
+module.exports = config;